Remove an obsolete TODO
The module-under-test cannot be compiled reliably from doctest.erl, as
it might need a special include path, or parse_transform, or something
else. Thus, it is up to the user to ensure that the code path defined in
the escript can find the file. In the future, we should allow the code
path to be set dynamically.